Made TabelAlbums its own component
