+++ client/src/components/TableAlbums/TableAlbums.js
@@ -1,72 +1,47 @@
-import React, { Component, Fragment } from 'react';
-import { connect } from 'react-redux';
+import React, { Component } from 'react';
+import ReactTable from 'react-table';
 
-import Loading from '../Loading';
-import Table from '../Table';
-
-class TableAlbums extends Component {
-  renderHeader = () => {
-    return (
-      <thead>
-        <tr>
-          <th />
-          <th>Name</th>
-          <th>Label</th>
-        </tr>
-      </thead>
-    );
-  };
-
-  renderBody = () => {
-    const { data } = this.props;
-
-    return (
-      <tbody>
-        {data.map(item => {
-          if (item.type === 'album') {
-            const { _id, label, name } = item;
-
-            return (
-              <tr key={_id}>
-                <td />
-                <td>{name}</td>
-                <td>{label}</td>
-              </tr>
-            );
-          }
-          return null;
-        })}
-      </tbody>
-    );
-  };
+export default class TableAlbums extends Component {
   render() {
-    const { props, renderBody, renderHeader } = this;
-    const { loading, data } = props;
+    const { albums, loading, onRowClick } = this.props;
+
+    const columns = [
+      {
+        Header: 'Name',
+        accessor: 'name',
+      },
+      {
+        Header: 'Number of Tracks',
+        accessor: 'number_of_tracks',
+      },
+      {
+        Header: 'Label',
+        accessor: 'label',
+      },
+      {
+        Header: 'Updated At',
+        accessor: 'updated_at',
+        Cell: row => {
+          let dateObj = new Date(row.value);
+          return (
+            dateObj.toLocaleDateString() + ' ' + dateObj.toLocaleTimeString()
+          );
+        },
+      },
+    ];
 
     return (
-      <Fragment>
-        {loading && <Loading />}
-        {data.length > 0 ? (
-          <Table>
-            {renderHeader()}
-            {renderBody()}
-          </Table>
-        ) : null}
-      </Fragment>
+      <ReactTable
+        className="-highlight clickable-rows"
+        columns={columns}
+        data={albums}
+        loading={loading}
+        showPageSizeOptions={false}
+        defaultPageSize={100}
+        minRows={3} // so the formatting does not look weird when there are 0 records
+        noDataText="This artist does not have any albums"
+        getTdProps={onRowClick}
+      />
     );
   }
 }
-
-function mapStateToProps(state) {
-  const { data, loading } = state.search;
-
-  return {
-    data,
-    loading,
-  };
-}
-
-export default connect(
-  mapStateToProps,
-  null,
-)(TableAlbums);
