+++ client/src/pages/ArtistViewPage/_ArtistViewPage.scss
@@ -1,5 +1,5 @@
 .artist-view-page {
-  &__album-table {
+  &__album-table-container {
     margin-top: 1rem;
   }
 }
